Added more unit tests, adapted existing code to new COntainerSlot standards

+++ ContainerSlot.cpp
@@ -54,12 +54,12 @@ std::string EmptySlot::toString(){
 /**
  * Container Constructor
 */
-Container::Container(std::string name, float mass, int x, int y, Origin o) : ContainerSlot(name, x, y, mass, true, false, o){
+Container::Container(std::string name, int mass, int x, int y, Origin o) : ContainerSlot(name, x, y, mass, true, false, o){
     
     //  Standard checks here
     if(name.empty() || name == "NAN" || name == "UNUSED"){throw std::invalid_argument("Invalid Container name: " + name);}; //  Check for invalid names
     if(name.length() > 256){throw std::invalid_argument("Container name too long.");} //  Name length
-    if(mass < 0 || mass > 99999.0){ throw std::invalid_argument("Invalid mass for container: " + std::to_string(mass));} //  Mass number
+    if(mass < 0 || mass > 99999){ throw std::invalid_argument("Invalid mass for container: " + std::to_string(mass));} //  Mass number
 };
 
 //  Container changing positions/Mass
